Fix status effect process overrides (#3087)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Fix void corruption and abrogation overriding process instead of using
tick.
Change messages in void corruption to first person.

This fixes void corruption being a super virus that can infect the whole
town in minutes due to taking 30x longer to tick.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
fix: Fixed void corruption tick interval, it will not spread to the
whole town in minutes anymore.
fix: Fixed void corruption never getting removed(?) Because it overrode
the duration tick.
fix: Fixed void corruption not having an alert icon.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

+++ code/modules/spells/spell_types/pointed/aoe/abrogation.dm
@@ -64,6 +64,7 @@
 	examine_text = "<b>SUBJECTPRONOUN is wreathed in a wild frenzy of ghostly motes!</b>"
 	effectedstats = list(STATKEY_STR = -2, STATKEY_CON = -2, STATKEY_END = -2, STATKEY_SPD = -2)
 	status_type = STATUS_EFFECT_REFRESH
+	tick_interval = 2 DECISECONDS
 	var/datum/weakref/debuffer
 	var/outline_colour = "#33cabc"
 	var/base_tick = 0.2
@@ -91,18 +92,20 @@
 	intensity += 1
 	to_chat(owner, span_boldwarning("The mists intensify, the glowing wisps steadily disrupting my body..."))
 
-/datum/status_effect/debuff/abrogation/process()
-	. = ..()
+/datum/status_effect/debuff/abrogation/tick()
 	if(!owner)
 		return
+
 	if(prob(66))
 		owner.adjustFireLoss(base_tick * intensity)
+
 	if(prob(10))
 		to_chat(owner, span_warning("A frenzy of ghostly motes assail my form!"))
 		owner.emote("scream")
 
+	if(!debuffer)
+		return
+
 	var/mob/living/our_debuffer = debuffer.resolve()
 	if(get_dist(our_debuffer, owner) > range)
 		qdel(src)
-
-
